+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/Modules.scala
@@ -90,6 +90,10 @@ import io.circe.ParsingFailure
 import io.circe.DecodingFailure
 import io.iohk.atala.agent.walletapi.sql.JdbcDIDSecretStorage
 import io.iohk.atala.resolvers.DIDResolver
+import io.iohk.atala.agent.walletapi.storage.DIDSecretStorage
+import io.iohk.atala.pollux.vc.jwt.{DidResolver => JwtDidResolver}
+import io.iohk.atala.castor.core.model.error.DIDOperationError.TooManyDidServiceAccess
+import io.iohk.atala.pollux.vc.jwt.PrismDidResolver
 
 object Modules {
 
@@ -152,12 +156,16 @@ object Modules {
     Server.start(port, app)
   }
 
-  val didCommExchangesJob: RIO[DIDResolver & HttpClient & CredentialService & ManagedDIDService, Unit] =
+  val didCommExchangesJob: RIO[
+    DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
+    Unit
+  ] =
     BackgroundJobs.didCommExchanges
       .repeat(Schedule.spaced(10.seconds))
       .unit
 
-  val presentProofExchangeJob: RIO[DIDResolver & HttpClient & PresentationService & ManagedDIDService, Unit] =
+  val presentProofExchangeJob
+      : RIO[DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService, Unit] =
     BackgroundJobs.presentProofExchanges
       .repeat(Schedule.spaced(10.seconds))
       .unit
@@ -395,6 +403,9 @@ object SystemModule {
 object AppModule {
   val didOpValidatorLayer: ULayer[DIDOperationValidator] = DIDOperationValidator.layer()
 
+  val didJwtResolverlayer: URLayer[DIDService, JwtDidResolver] =
+    ZLayer.fromFunction(PrismDidResolver(_))
+
   val didServiceLayer: TaskLayer[DIDService] =
     (didOpValidatorLayer ++ GrpcModule.layers) >>> DIDServiceImpl.layer
 
@@ -466,7 +477,8 @@ object HttpModule {
     (apiServiceLayer ++ apiMarshallerLayer) >>> ZLayer.fromFunction(new DIDRegistrarApi(_, _))
   }
 
-  val issueCredentialsProtocolApiLayer: RLayer[DidComm & ManagedDIDService & AppConfig, IssueCredentialsProtocolApi] = {
+  val issueCredentialsProtocolApiLayer
+      : RLayer[DidComm & ManagedDIDService & ConnectionService & AppConfig, IssueCredentialsProtocolApi] = {
     val serviceLayer = AppModule.credentialServiceLayer
     val apiServiceLayer = serviceLayer >>> IssueCredentialsProtocolApiServiceImpl.layer
     val apiMarshallerLayer = IssueCredentialsProtocolApiMarshallerImpl.layer