+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/jobs/BackgroundJobs.scala
@@ -7,10 +7,10 @@ import io.iohk.atala.pollux.core.model.IssueCredentialRecord
 import io.iohk.atala.pollux.core.model.error.CredentialServiceError
 import io.iohk.atala.pollux.core.service.CredentialService
 import io.iohk.atala.pollux.vc.jwt.W3cCredentialPayload
-import zio.*
-
+import zio.Duration
 import java.time.Instant
-
+import java.time.Clock
+import java.time.ZoneId
 import io.iohk.atala.mercury.DidComm
 import io.iohk.atala.mercury.MediaTypes
 import io.iohk.atala.mercury.MessagingService
@@ -22,7 +22,7 @@ import io.iohk.atala.mercury.protocol.presentproof.RequestPresentation
 import io.iohk.atala.resolvers.DIDResolver
 import io.iohk.atala.resolvers.UniversalDidResolver
 import java.io.IOException
-
+import io.iohk.atala.pollux.vc.jwt._
 import zhttp.service._
 import zhttp.http._
 import io.iohk.atala.pollux.vc.jwt.W3CCredential
@@ -34,6 +34,28 @@ import io.iohk.atala.agent.walletapi.service.ManagedDIDService
 import io.iohk.atala.mercury.AgentServiceAny
 import org.didcommx.didcomm.DIDComm
 import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError
+import io.iohk.atala.agent.walletapi.model.ManagedDIDTemplate
+import io.iohk.atala.agent.walletapi.sql.JdbcDIDSecretStorage
+import io.iohk.atala.pollux.vc.jwt.ES256KSigner
+import io.iohk.atala.castor.core.model.did._
+import java.security.KeyFactory
+import java.security.spec.EncodedKeySpec
+import java.security.spec.ECPrivateKeySpec
+import org.bouncycastle.jce.spec.ECNamedCurveSpec
+import org.bouncycastle.jce.ECNamedCurveTable
+import com.ionspin.kotlin.bignum.integer.BigInteger
+import com.ionspin.kotlin.bignum.integer.Sign
+import io.iohk.atala.pollux.vc.jwt.Issuer
+import java.security.spec.ECPublicKeySpec
+import java.security.spec.ECPoint
+import org.bouncycastle.jce.provider.BouncyCastleProvider
+import io.circe.Json
+import io.circe.syntax._
+import io.iohk.atala.agent.walletapi.storage.DIDSecretStorage
+import io.iohk.atala.agent.walletapi.model.error.CreateManagedDIDError
+import io.iohk.atala.pollux.vc.jwt.JWT
+import io.iohk.atala.pollux.vc.jwt.{DidResolver => JwtDidResolver}
+import io.iohk.atala.agent.walletapi.model._
 
 object BackgroundJobs {
 
@@ -58,7 +80,10 @@ object BackgroundJobs {
 
   private[this] def performExchange(
       record: IssueCredentialRecord
-  ): URIO[DIDResolver & HttpClient & CredentialService & ManagedDIDService, Unit] = {
+  ): URIO[
+    DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
+    Unit
+  ] = {
     import IssueCredentialRecord._
     import IssueCredentialRecord.ProtocolState._
     import IssueCredentialRecord.PublicationState._
@@ -132,7 +157,8 @@ object BackgroundJobs {
           // Set PublicationState to PublicationPending
           for {
             credentialService <- ZIO.service[CredentialService]
-            issuer = credentialService.createIssuer
+            // issuer = credentialService.createIssuer
+            issuer <- createPrismDIDIssuer()
             w3Credential <- credentialService.createCredentialPayloadFromRecord(
               record,
               issuer,
@@ -220,9 +246,58 @@ object BackgroundJobs {
       }
   }
 
+  // TODO: Improvements needed here:
+  // - A single PrismDID genrated at agent startup should be used.
+  // - For now, we include the long form in the JWT credential to facilitate validation on client-side, but resolution should be used instead.
+  // - There should be a way to retrieve the 'default' PrismDID from ManagedDIDService (use of an alias in DB record?)
+  // - Simplify convertion of ECKeyPair to JDK security classes
+  // - ECPrivateKey should probably remain 'private' and signing operation occur in ManagedDIDService
+  private[this] def createPrismDIDIssuer(): ZIO[ManagedDIDService & DIDSecretStorage, Throwable, Issuer] = {
+    for {
+      managedDIDService <- ZIO.service[ManagedDIDService]
+      longFormPrismDID <- managedDIDService.createAndStoreDID(
+        ManagedDIDTemplate(
+          Seq(
+            DIDPublicKeyTemplate("issuing", VerificationRelationship.Authentication)
+          ),
+          Nil
+        )
+      )
+      didSecretStorage <- ZIO.service[DIDSecretStorage]
+      maybeECKeyPair <- didSecretStorage.getKey(longFormPrismDID.asCanonical, "issuing")
+      _ <- ZIO.logInfo(s"ECKeyPair => $maybeECKeyPair")
+      maybeIssuer <- ZIO.succeed(maybeECKeyPair.map(ecKeyPair => {
+        val ba = ecKeyPair.privateKey.toPaddedByteArray(EllipticCurve.SECP256K1)
+        val keyFactory = KeyFactory.getInstance("EC", new BouncyCastleProvider())
+        val ecParameterSpec = ECNamedCurveTable.getParameterSpec("secp256k1")
+        val ecNamedCurveSpec = ECNamedCurveSpec(
+          ecParameterSpec.getName(),
+          ecParameterSpec.getCurve(),
+          ecParameterSpec.getG(),
+          ecParameterSpec.getN()
+        )
+        val ecPrivateKeySpec = ECPrivateKeySpec(java.math.BigInteger(1, ba), ecNamedCurveSpec)
+        val privateKey = keyFactory.generatePrivate(ecPrivateKeySpec)
+        val bcECPoint = ecParameterSpec
+          .getG()
+          .multiply(privateKey.asInstanceOf[org.bouncycastle.jce.interfaces.ECPrivateKey].getD())
+        val ecPublicKeySpec = ECPublicKeySpec(
+          new ECPoint(
+            bcECPoint.normalize().getAffineXCoord().toBigInteger(),
+            bcECPoint.normalize().getAffineYCoord().toBigInteger()
+          ),
+          ecNamedCurveSpec
+        )
+        val publicKey = keyFactory.generatePublic(ecPublicKeySpec)
+        Issuer(io.iohk.atala.pollux.vc.jwt.DID(longFormPrismDID.toString), ES256KSigner(privateKey), publicKey)
+      }))
+      issuer = maybeIssuer.get
+    } yield issuer
+  }
+
   private[this] def performPresentation(
       record: PresentationRecord
-  ): URIO[DIDResolver & HttpClient & PresentationService & ManagedDIDService, Unit] = {
+  ): URIO[DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService, Unit] = {
     import io.iohk.atala.pollux.core.model.PresentationRecord.ProtocolState._
 
     val aux = for {
@@ -270,14 +345,43 @@ object BackgroundJobs {
               } yield ()
         case PresentationRecord(id, _, _, _, _, _, _, _, PresentationSent, _, _, _) =>
           ZIO.logDebug("PresentationRecord: PresentationSent") *> ZIO.unit
-        case PresentationRecord(id, _, _, _, _, _, _, _, PresentationReceived, _, _, _) =>
+        case PresentationRecord(id, _, _, _, _, _, _, _, PresentationReceived, _, _, presentation) => // Verifier
           ZIO.logDebug("PresentationRecord: PresentationReceived") *> ZIO.unit
-          for {
-            _ <- ZIO.log(s"PresentationRecord: PresentationPending (Send Massage)")
-            // TODO Verify  https://input-output.atlassian.net/browse/ATL-2702
-            service <- ZIO.service[PresentationService]
-            _ <- service.markPresentationVerified(id)
-          } yield ()
+          val clock = java.time.Clock.system(ZoneId.systemDefault)
+          presentation match
+            case None => ZIO.fail(InvalidState("PresentationRecord in 'PresentationReceived' with no Presentation"))
+            case Some(p) =>
+              for {
+                _ <- ZIO.log(s"PresentationRecord: 'PresentationReceived' ")
+                didResolverService <- ZIO.service[JwtDidResolver]
+                credentialsValidationResult <- p.attachments.head.data match {
+                  case Base64(data) =>
+                    val base64Decoded = new String(java.util.Base64.getDecoder().decode(data)).drop(1).dropRight(1)
+
+                    println(s"Base64decode:\n\n ${base64Decoded} \n\n")
+                    JwtPresentation.verify(
+                      JWT(base64Decoded),
+                      JwtPresentation.PresentationVerificationOptions(
+                        maybeProofPurpose = Some(VerificationRelationship.Authentication),
+                        verifySignature = false,
+                        verifyDates = true,
+                        leeway = Duration.Zero,
+                        maybeCredentialOptions = Some(
+                          CredentialVerification.CredentialVerificationOptions(
+                            verifySignature = true,
+                            verifyDates = false,
+                            leeway = Duration.Zero,
+                            maybeProofPurpose = Some(VerificationRelationship.Authentication)
+                          )
+                        )
+                      )
+                    )(didResolverService)(clock)
+                  case any => ZIO.fail(NotImplemented)
+                }
+                _ <- ZIO.log(s"CredentialsValidationResult: $credentialsValidationResult")
+                service <- ZIO.service[PresentationService]
+                _ <- service.markPresentationVerified(id)
+              } yield ()
         // TODO move the state to PresentationVerified
         case PresentationRecord(id, _, _, _, _, _, _, _, PresentationVerified, _, _, _) =>
           ZIO.logDebug("PresentationRecord: PresentationVerified") *> ZIO.unit